requirements.txt
PySide6~=6.3.1
setuptools~=58.1.0
Pillow~=9.1.1
requests~=2.27.1
oss2~=2.16.0
python-dateutil~=2.8.2
pygame~=2.1.2
pandas~=1.4.2
numpy~=1.22.3